[SPARK-51747][SQL][FOLLOW-UP] Data source cached plan conf and migration guide
### What changes were proposed in this pull request?

Follow-up to https://github.com/apache/spark/pull/50538.

Add a SQL legacy conf to enable/disable the change to allow users to restore the previous behavior. Also add a migration guide note.

### Why are the changes needed?

The original PR changes the behavior of reading from a data source file with options. The flag is needed to allow users a way to restore the former behavior, if desired.

### Does this PR introduce _any_ user-facing change?

No (original PR was a user-facing change, but this PR simply adds a config).

### How was this patch tested?

Added test for the config

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #50571 from asl3/asl3/filedatasourcecache-docsconf.

Authored-by: Amanda Liu <amanda.liu@databricks.com>
Signed-off-by: Gengliang Wang <gengliang@apache.org>