[SPARK-51800][INFRA] Set up the CI for UDS in PySpark
### What changes were proposed in this pull request?

This PR is a followup of https://github.com/apache/spark/pull/50466 that sets the CI for UDS in Python.

### Why are the changes needed?

It might be flaky to make a mistake in the codebase. We should refactor a bit as well to prevent this.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Setting the CI for it. Will monitor the build.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50585 from HyukjinKwon/SPARK-51800.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>