[SPARK-51414][SQL] Add the make_time() function
### What changes were proposed in this pull request?

This adds a new function `make_time` for the new time datatype. A new function `MakeTime` is added and some logic was refactored out of `MakeTimestamp` to be shared with the new function.

### Why are the changes needed?

This is needed to create time data type objects.

### Does this PR introduce _any_ user-facing change?

Yes, a new function `make_time` is added.

### How was this patch tested?

New unit tests and manually spark-shell testing.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #50269 from robreeves/SPARK-51414.

Lead-authored-by: Rob Reeves <roreeves@linkedin.com>
Co-authored-by: Rob Reeves <robertpreeves@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>