[SPARK-51814][SS][PYTHON] Introduce a new API `transformWithState` in PySpark
### What changes were proposed in this pull request?

This PR proposes to introduce a new API named `transformWithState` in PySpark, which uses PySpark Row type as the type of input and return type for user function. Note that this is not intended to replace the existing API `transformWithStateInPandas`, since users may rely on the feature of vectorization and also libraries which depend on Pandas.

The new API is quite same with existing `transformWithStateInPandas`, except the method name and the signature of user functions. The difference of signature of user functions are also quite straightforward - replace Pandas with Row.

There are several points on reviewing:

1. This PR tries to minimize the code diff via refactoring existing classes and reusing them. These classes are mostly renamed from Pandas to PySpark, because it's no longer used only for Pandas. Spark will pass user facing type as a parameter for internal execution classes to distinguish Pandas vs Row execution.
2. Same applies to config, hence this PR renames the existing config from `spark.sql.execution.arrow.transformWithStateInPandas.maxRecordsPerBatch` to `spark.sql.execution.arrow.transformWithStateInPySpark.maxRecordsPerBatch`. This PR retains the support of old name via `withAlternative` to not break backward compatibility.
3. From 1, `TransformWithStateInPandasExec` is renamed to `TransformWithStateInPySparkExec`. Since this affects both state operation metadata and state store reader, this PR retains the short name of physical node to `transformWithStateInPandasExec` when user facing type is set to Pandas.
4. Python test is majorly refactored to allow testing both Pandas and Row. Regardless of the type, the query is expected to produce the same output as long as we implement the stateful processor for the same behavior.

This PR does not address the support of Spark Connect - while it is easy and straightforward to support, we have to modify the proto which needs separate eyes.

### Why are the changes needed?

We realized the overhead on converting Arrow RecordBatch to Pandas DataFrame is non-trivial, especially the case where cardinality is quite high (e.g. every data has its unique key). This case we see more than 10 times slower than the case of low cardinality. Using Row reduces a lot on the overhead of conversion, which addresses the huge gap of performance.

(We considered Arrow as a type since there are several pairs of Arrow <-> Pandas for existing PySpark API. We rejected the approach because we don't see a strong evidence that Arrow type of API is being widely used, and PySpark Row is lot easier to use.)

### Does this PR introduce _any_ user-facing change?

Yes, this PR introduces a new user facing API.

### How was this patch tested?

Tests are refactored to verify the new API.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50600 from HeartSaVioR/SPARK-51814.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>