[SPARK-51905][SQL] Disallow NOT ENFORCED CHECK constraint
### What changes were proposed in this pull request?

As we plan to enforce check constraints by default, it remains unclear what the behavior should be when it is "NOT ENFORCED".

- Option 1: Never support NOVALIDATE/NOT VALID in ALTER TABLE. Use ENFORCED/NOT ENFORCED to check if validation must be done.
- Option 2: Add NOVALIDATE/NOT VALID to control this.
- Option 3: Don't support of "NOT ENFORCED" for now. Let's think about Option 1 or Option 2 later

I would prefer option 3. We can make a decision later based on the user expectations.

### Why are the changes needed?

Clarify the behavior of a "NOT ENFORCED" CHECK constraint.

### Does this PR introduce _any_ user-facing change?

No, this is DSV2 framework and it is not released yet.
### How was this patch tested?

New UT

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #50772 from gengliangwang/disallowNotEnforcedCheck.

Authored-by: Gengliang Wang <gengliang@apache.org>
Signed-off-by: Gengliang Wang <gengliang@apache.org>