[SPARK-52037][SQL] Order in LCA inner project lists should respect order in original project list
### What changes were proposed in this pull request?
Enforce that the order in inner project lists generated when resolving LCAs respects ordering of aliases in original project list.

For a query:
```SELECT 0 AS a, 1 AS b, 2 AS c, b AS d, a AS e, d AS f, a AS g, g AS h, h AS i```

Before this change, plan would be:

```
Project [a, b, 2 AS c, d, a AS e, d AS f, g, h, h AS i]
+- Project [b, a, d, g, g AS h]
     +- Project [b, a, b AS d, a AS g]
          +- Project [1 AS b, 0 AS a]
               +- OneRowRelation
```

After this change, `a` should come before `b` in all inner `Project`s:

```
Project [a, b, 2 AS c, d, a AS e, d AS f, g, h, h AS i]
+- Project [a, b, d, g, g AS h]
     +- Project [a, b, b AS d, a AS g]
          +- Project [0 AS a, 1 AS b]
               +- OneRowRelation
```

### Why are the changes needed?
This change is necessary because single-pass analyzer always respects the order from the original project list there is a logical plan mismatch for the example query.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
Added a test case.

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #50828 from mihailotim-db/mihailotim-db/lca_ordering.

Authored-by: Mihailo Timotic <mihailo.timotic@databricks.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>