[SPARK-51981][SS] Add JobTags to queryStartedEvent
### What changes were proposed in this pull request?
Adding a new jobTags parameter for QueryStartedEvent so that it can be connected to the actual spark connect command that triggered this streaming. Also besides adding the parameter, a fix has been applied to the timestamp because previously json reads the wrong argument

### Why are the changes needed?
Without this, there is no way to tell where does this streaming originate from.

### Does this PR introduce _any_ user-facing change?

No

### How was this patch tested?

Unit test is added

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #50780 from gjxdxh/lingkai-kong_data/SPARK-51981.

Authored-by: Lingkai Kong <lingkai.kong@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>