[SPARK-52129] Improve ParserUtils to support more Scripting constructs
### What changes were proposed in this pull request?

#### Label casing fix
In `enterLabeledScope()` the original name was used instead of the lowercased one. This caused issues in some cases when matching labels.

#### CompoundBodyParsingContext
Renamed previous `SqlScriptingParsingContext` to a more specific `CompoundBodyParsingContext` to free up the name for a more generic context.

#### Introduced new SqlScriptingParsingContext
It consists of previously existing `SqlScriptingLabelContext` and newly introduced `SqlScriptingConditionContext` for the sake of tracking the declared condition names. This is required in order to be able to detect duplicate condition names and map specific conditions to their assigned SQL states.

#### Other changes
Propagating the changed structures through AstBuilder.
Renaming error condition and related functions.

### Why are the changes needed?

These changes introduce support for new correctness checks during parsing time for SQL scripts.
The checks and changes are explained in the previous section.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

Existing and new unit tests.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50893 from davidm-db/parsing_contexts.

Lead-authored-by: David Milicevic <david.milicevic@databricks.com>
Co-authored-by: David Milicevic <163021185+davidm-db@users.noreply.github.com>
Signed-off-by: Wenchen Fan <wenchen@databricks.com>