[SPARK-52131][INFRA] Upload the application directory when tests for the `yarn` module fail
### What changes were proposed in this pull request?
This pr adds functionality to upload the contents of the `target/test/data/` directory when the `yarn` module tests fail in GitHub Actions. This directory stores information related to Yarn Applications, such as logs and job configurations. The upload of this directory's contents facilitates troubleshooting when `yarn` module tests encounter failures.

### Why are the changes needed?
This facilitates troubleshooting when tests for the `yarn` module fail.

### Does this PR introduce _any_ user-facing change?
No

### How was this patch tested?
- Pass GitHub Actions
- It has been confirmed that the expected files can be downloaded when test failures occur in the `yarn` module.
https://github.com/LuciferYang/spark/actions/runs/15019569409

![image](https://github.com/user-attachments/assets/fd22f7d9-64fd-4d6f-a0d8-5e13c70de272)

After unzipping, the contents of the folder are as follows:

![image](https://github.com/user-attachments/assets/a2dce30f-add2-4671-9f64-2c26a25a8bab)

### Was this patch authored or co-authored using generative AI tooling?
No

Closes #50891 from LuciferYang/upload-yarn-app-log.

Lead-authored-by: yangjie01 <yangjie01@baidu.com>
Co-authored-by: YangJie <yangjie01@baidu.com>
Signed-off-by: yangjie01 <yangjie01@baidu.com>