[SPARK-52127][SS][PYTHON] Squeeze the protocol of MapState KEYS / VALUES / ITERATOR for transformWithState in PySpark
### What changes were proposed in this pull request?

This PR proposes to squeeze the protocol of MapState KEYS / VALUES / ITERATOR for transformWithState in PySpark, which will help a lot on dealing with small map on MapState.

Here are the changes:

* MapState.keys(), MapState.values(), MapState.iterator() no longer requires additional request to notice there is no further data to read.
  * We inline the data into proto message, to ease of determine whether the iterator has fully consumed or not.

This change is the same mechanism we applied for ListState.get(). We got performance improvement in the prior case, and we also see this change to be helpful on our internal benchmark.

### Why are the changes needed?

To optimize further on MapState operations.

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?

New UT.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #50885 from HeartSaVioR/SPARK-52127.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>