[SPARK-52241][BUILD][FOLLOW-UP] Make it fail if one of the curls fails
### What changes were proposed in this pull request?

This PR is a followup of https://github.com/apache/spark/pull/50965 that fails if one of the curls fails.

### Why are the changes needed?

This is a regression in the release script. We should make it failed if one of the uploads fails.

### Does this PR introduce _any_ user-facing change?

No, dev-only.

### How was this patch tested?

Manually tested.

### Was this patch authored or co-authored using generative AI tooling?

No

Closes #50968 from HyukjinKwon/SPARK-52241-followup.

Authored-by: Hyukjin Kwon <gurwls223@apache.org>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>
(cherry picked from commit 41434ceb1dabea717e0190e308fefd78379df87f)
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>