[SPARK-52324][INFRA] Move Spark docs to the release directory
### What changes were proposed in this pull request?

The docs of old Spark versions are served from the Apache release directory instead of the spark-website. We need to move Spark docs to the Apache release directory when finalizing the release.

This PR updates the release scripts to do so.

### Why are the changes needed?

release process

### Does this PR introduce _any_ user-facing change?

no

### How was this patch tested?

No, we will test it when finalizing Spark 3.5.6

### Was this patch authored or co-authored using generative AI tooling?

no

Closes #51026 from cloud-fan/release.

Authored-by: Wenchen Fan <wenchen@databricks.com>
Signed-off-by: Hyukjin Kwon <gurwls223@apache.org>