[SPARK-50752][SQL][PYSPARK][FOLLOWUP] Respect PYTHON_UDF_MAX_RECORDS_PER_BATCH in Python worker
### What changes were proposed in this pull request?

This PR is a follow-up of SPARK-50752 (PR #49397), which introduced PYTHON_UDF_MAX_RECORDS_PER_BATCH into SQLConf to control the batch size of normal Python UDF.

This PR is to fix the config to be effective for Python worker side as well.

### Why are the changes needed?

The original PR enabled the control of batch size for JVM side, but it wasn't properly propagated to Python worker since we missed to override the value for Python UDF. (Default value before overriding is a static one, 100)

### Does this PR introduce _any_ user-facing change?

Yes, but it's about tuning and does not impact any change on the output.

### How was this patch tested?

No automated test, since there is no good way to test this properly since missing this does not change the output itself.

This was manually tested with artificial debug logging.

### Was this patch authored or co-authored using generative AI tooling?

No.

Closes #51124 from HeartSaVioR/SPARK-50752-FOLLOWUP-respect-PYTHON_UDF_MAX_RECORDS_PER_BATCH-in-python-worker.

Authored-by: Jungtaek Lim <kabhwan.opensource@gmail.com>
Signed-off-by: Jungtaek Lim <kabhwan.opensource@gmail.com>