[SPARK-52460][SQL] Store internal TIME values as nanoseconds
### What changes were proposed in this pull request?
In the PR, I propose to store internal TIME values as nanoseconds in `Long` since the midnight instead of microseconds in `Long`.

### Why are the changes needed?
`Long` with nanoseconds precision can keep the full range of TIME values from 0 to 24 * 60 * 60 * 1000 * 1000 * 1000 - 1 which is less than maximum of Long. This will simplify support of `TIME(9)`, for example.

### Does this PR introduce _any_ user-facing change?
No.

### How was this patch tested?
By running the affected test suites:
```
$ build/sbt "sql/testOnly org.apache.spark.sql.SQLQueryTestSuite"
$ build/sbt "test:testOnly *TimeExpressionsSuite"
$ build/sbt "test:testOnly *TimeFormatterSuite"
```

### Was this patch authored or co-authored using generative AI tooling?
No.

Closes #51156 from MaxGekk/time-nanos.

Authored-by: Max Gekk <max.gekk@gmail.com>
Signed-off-by: Max Gekk <max.gekk@gmail.com>